updayed setup_env

+++ setup_env.sh
@@ -1,5 +1,8 @@
 #!/bin/env bash
 FILENAME=".env"
-rm $FILENAME
+if test -f "$FILENAME"; then
+   rm $FILENAME
+fi
+
 touch $FILENAME
 echo REACT_APP_API_URL=https://rugged-badlands-76158.herokuapp.com >> $FILENAME
